+++ src/Pim/Component/Catalog/Completeness/Checker/Attribute/MetricCompleteChecker.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Completeness\Checker\Attribute;
+namespace Pim\Component\Catalog\Completeness\Checker\Attribute;
 
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\ChannelInterface;