+++ src/Pim/Component/Catalog/Completeness/Checker/CompleteCheckerRegistryInterface.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Pim\Component\Catalog\Completeness\Checker;
+
+use Pim\Component\Catalog\Completeness\Checker\Attribute\AttributeCompleteCheckerInterface;
+
+/**
+ * @author    JM Leroux <jean-marie.leroux@akeneo.com>
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+interface CompleteCheckerRegistryInterface
+{
+    /**
+     * @return AttributeCompleteCheckerInterface[]
+     */
+    public function getAttributeCheckers();
+
+    /**
+     * @param AttributeCompleteCheckerInterface $attributeCompleteChecker
+     */
+    public function registerAttributeChecker(AttributeCompleteCheckerInterface $attributeCompleteChecker);
+}